TMLEC_QRScan/model_setup/requirements.txt
tensorflow>=2.12.0
onnx>=1.13.0
onnxruntime>=1.14.0
onnx-tf>=1.10.0
tf2onnx>=1.13.0
numpy>=1.24.3
opencv-python>=4.7.0
pillow>=9.5.0 